backend/requirements.txt
openai
faiss-cpu
beautifulsoup4
tqdm
sentence-transformers   # optional fallback
numpy
python-dotenv
